Use requestAnimationFrame polyfill to avoid test warning
This gets rid of a test warning by providing a requestAnimationFrame polyfill
before the tests are run. This approach was suggested in an issue [here](https://github.com/facebook/jest/issues/4545#issuecomment-332762365)

The workaround to the `jest` config inside of package.json is to use the
`setupTestFrameworkScriptFile` flag when running jest.